:sparkles: Alterações nas classes Oportunidade Controller, Oportunidade (Model), Oportunidade Repository e Oportunidade Service

+++ src/main/java/com/generation/crm_backend/service/OportunidadeService.java
@@ -8,15 +8,25 @@ import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 import org.springframework.web.server.ResponseStatusException;
 
+import com.generation.crm_backend.model.Cliente;
 import com.generation.crm_backend.model.Oportunidade;
 import com.generation.crm_backend.model.StatusOportunidade;
+import com.generation.crm_backend.model.Usuario;
+import com.generation.crm_backend.repository.ClienteRepository;
 import com.generation.crm_backend.repository.OportunidadeRepository;
+import com.generation.crm_backend.repository.UsuarioRepository;
 
 @Service
 public class OportunidadeService {
 
 	@Autowired
 	private OportunidadeRepository oportunidadeRepository;
+	
+	@Autowired
+    private ClienteRepository clienteRepository; //injeta clienteRepository
+
+	@Autowired
+	private UsuarioRepository usuarioRepository; //injeta usuarioRepository
 
 	public List<Oportunidade> findAll() {
 		return oportunidadeRepository.findAll();
@@ -30,30 +40,59 @@ public class OportunidadeService {
 		return oportunidadeRepository.findByStatus(status);
 	}
 
-	public List<Oportunidade> findByIdCliente(Long idCliente) {
-		return oportunidadeRepository.findByIdCliente(idCliente);
+	public List<Oportunidade> findByClienteId(Long idCliente) {
+		return oportunidadeRepository.findByClienteId(idCliente);
 	}
 
-	public List<Oportunidade> findByIdUsuario(Long idUsuario) {
-		return oportunidadeRepository.findByIdUsuario(idUsuario);
+	public List<Oportunidade> findByUsuarioId(Long idUsuario) {
+		return oportunidadeRepository.findByUsuarioId(idUsuario);
 	}
 
+	public List<Oportunidade> findAllByTituloContainingIgnoreCase(String titulo) {
+		return oportunidadeRepository.findAllByTituloContainingIgnoreCase(titulo);
+	}
+
+	//lidar com objetos cliente usuario
 	public Oportunidade save(Oportunidade oportunidade) {
+		Cliente cliente = clienteRepository.findById(oportunidade.getCliente().getId())
+				.orElseThrow(() -> new ResponseStatusException
+						(HttpStatus.NOT_FOUND, "Cliente não encontrado!"));
+
+		Usuario usuario = usuarioRepository.findById(oportunidade.getUsuario().getId())
+				.orElseThrow(() -> new ResponseStatusException
+						(HttpStatus.NOT_FOUND, "Usuário não encontrado!"));
+
+		oportunidade.setCliente(cliente);
+		oportunidade.setUsuario(usuario);
+
 		return oportunidadeRepository.save(oportunidade);
 	}
 
+	//lidar com obj cliente usuario
 	public Oportunidade update(Oportunidade oportunidade) {
 
 		if (oportunidade.getId() == null || !oportunidadeRepository.existsById(oportunidade.getId())) {
-
 			throw new ResponseStatusException(HttpStatus.NOT_FOUND,
 					"Oportunidade com ID " + oportunidade.getId() + " não encontrada para atualização...");
 		}
+
+		Cliente cliente = clienteRepository.findById(oportunidade.getCliente().getId())
+				.orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, "Cliente não encontrado!"));
+
+		Usuario usuario = usuarioRepository.findById(oportunidade.getUsuario().getId())
+				.orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, "Usuário não encontrado!"));
+
+		oportunidade.setCliente(cliente);
+		oportunidade.setUsuario(usuario);
 
 		return oportunidadeRepository.save(oportunidade);
 	}
 
 	public void deleteById(Long id) {
+		if (!oportunidadeRepository.existsById(id)) {
+			throw new ResponseStatusException(HttpStatus.NOT_FOUND,
+					"Oportunidade com ID " + id + " não encontrada para exclusão!");
+		}
 		oportunidadeRepository.deleteById(id);
 	}
